moved tests to parser
