+++ tests/http_spec.rs
@@ -301,27 +301,3 @@ async fn run(spec: HttpSpec, downstream_assertion: &&DownstreamAssertion) -> any
     handle_single_request(req, server_context).await
   }
 }
-
-#[cfg(test)]
-mod parser_tests {
-  use tailcall::async_graphql_hyper::GraphQLRequest;
-  use tailcall::parser::de::Parser;
-
-  #[test]
-  fn t1_url_qry_parser() {
-    let parser = Parser::from_path("api/user?$=name,age,address.city,address.state");
-    assert_eq!(
-      parser.unwrap().parse::<GraphQLRequest>().unwrap().0.query,
-      "{user {address {city state} age name}}"
-    );
-  }
-  #[test]
-  fn t2_url_nested_qry_parser() {
-    let parser =
-      Parser::from_path("api/user?id=123,address.country=India,address.city=Foo&$=name,age,address.city,address.state");
-    assert_eq!(
-      parser.unwrap().parse::<GraphQLRequest>().unwrap().0.query,
-      "{user (id: 123,) {address (city: Foo,country: India,) ) {city state} age name}}"
-    );
-  }
-}
